Servicios terminados

+++ WorkMerge/src/main/java/com/WorkMerge/services/AdminService.java
@@ -1,5 +1,9 @@
 package com.WorkMerge.services;
 
+
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -7,7 +11,11 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.WorkMerge.entities.Admin;
+import com.WorkMerge.entities.Client;
+import com.WorkMerge.entities.Company;
+import com.WorkMerge.entities.Job;
 import com.WorkMerge.enums.Rol;
+import com.WorkMerge.exceptions.ServiceException;
 import com.WorkMerge.repositories.AdminRepository;
 import com.WorkMerge.repositories.ClientRepository;
 import com.WorkMerge.repositories.CompanyRepository;
@@ -31,7 +39,9 @@ public class AdminService {
 	/*Crear ADMIN, Eliminar cliente, Eliminar empresa, Eliminar trabajo*/
 	
 	@Transactional(propagation = Propagation.REQUIRED, rollbackFor = { Exception.class })
-	public Admin newAdmin(String email, String password, String nickname) {
+	public Admin newAdmin(String email, String password, String nickname) throws ServiceException{
+		
+		validate(email, password, nickname);
 		
 		/*Encripto Contraseña*/
 		String encryptedKey = new BCryptPasswordEncoder().encode(password);
@@ -49,24 +59,66 @@ public class AdminService {
 	}
 	
 	@Transactional(propagation = Propagation.REQUIRED, rollbackFor = { Exception.class })
-	public void deleteClient(String id) {
+	public void deleteClient(String id) throws ServiceException{
 		
-		/*Elimino cliente*/
-		clientRepository.deleteById(id);
+		Optional<Client> respuesta = clientRepository.findById(id);
+		if(respuesta.isPresent()) {
+			/*Elimino cliente*/
+			clientRepository.deleteById(id);
+		} else {
+			throw new ServiceException("No se encontro el usuario a borrar.");
+		}
 	}
 	
 	@Transactional(propagation = Propagation.REQUIRED, rollbackFor = { Exception.class })
-	public void deleteCompany(String id) {
+	public void deleteCompany(String id) throws ServiceException{
 		
-		/*Elimino empresa*/
-		companyRepository.deleteById(id);
+		Optional<Company> respuesta = companyRepository.findById(id);
+		if(respuesta.isPresent()) {
+			/*Elimino empresa*/
+			companyRepository.deleteById(id);
+		} else {
+			throw new ServiceException("No se encontro la compañia a borrar.");
+		}
 	}
 	
 	@Transactional(propagation = Propagation.REQUIRED, rollbackFor = { Exception.class })
-	public void deleteJob(String id) {
+	public void deleteJob(String id) throws ServiceException{
 		
-		/*Elimino trabajo*/
-		jobRepository.deleteById(id);
+		Optional<Job> respuesta = jobRepository.findById(id);
+		if(respuesta.isPresent()) {
+			/*Elimino trabajo*/
+			jobRepository.deleteById(id);
+		} else {
+			throw new ServiceException("No se encontro el trabajo a borrar.");
+		}
+	}
+	
+	@Transactional(readOnly = true)
+	public List<Job> listJobs() {
+		return jobRepository.findAll();
+	}
+
+	@Transactional(readOnly = true)
+	public List<Company> listCompanies(){
+		return companyRepository.findAll();
+	}
+
+	@Transactional(readOnly = true)
+	public List<Client> listClients(){
+		return clientRepository.findAll();
+	}
+	
+	private void validate(String email, String password, String nickname) throws ServiceException{
+		if(email == null || email.isEmpty()) {
+			throw new ServiceException("El email no puede ser nulo o estar vacio.");
+		}
+		if(password == null || password.isEmpty()) {
+			throw new ServiceException("La contraseña no puede ser nula o estar vacia.");
+		}
+		if(nickname == null || nickname.isEmpty()) {
+			throw new ServiceException("El nick no puede ser nulo o estar vacio.");
+		}
 	}
 	
 	/*
